api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt base.in
#
alembic==1.13.1
    # via -r base.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
boto3==1.34.69
    # via -r base.in
botocore==1.34.69
    # via
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via uvicorn
datetime==5.5
    # via -r base.in
decouple==0.0.7
    # via -r base.in
deprecation==2.1.0
    # via postgrest
fastapi==0.110.0
    # via -r base.in
gotrue==2.4.1
    # via supabase
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httpx==0.25.2
    # via
    #   gotrue
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
idna==3.6
    # via
    #   anyio
    #   httpx
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mako==1.3.2
    # via alembic
markupsafe==2.1.5
    # via mako
packaging==24.0
    # via deprecation
postgrest==0.16.2
    # via supabase
pydantic==2.6.4
    # via
    #   -r base.in
    #   fastapi
    #   gotrue
    #   postgrest
pydantic-core==2.16.3
    # via pydantic
pyjwt==2.8.0
    # via -r base.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   realtime
    #   storage3
python-multipart==0.0.9
    # via -r base.in
pytz==2024.1
    # via datetime
realtime==1.0.2
    # via supabase
s3transfer==0.10.1
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.29
    # via
    #   -r base.in
    #   alembic
starlette==0.36.3
    # via fastapi
storage3==0.7.4
    # via supabase
strenum==0.4.15
    # via postgrest
supabase==2.4.0
    # via -r base.in
supafunc==0.3.3
    # via supabase
typing-extensions==4.10.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   realtime
    #   sqlalchemy
    #   storage3
urllib3==2.2.1
    # via botocore
uvicorn==0.29.0
    # via -r base.in
websockets==11.0.3
    # via realtime
zope-interface==6.2
    # via datetime

# The following packages are considered to be unsafe in a requirements file:
# setuptools
